refactor: rename view rect files and renderer API
