+++ view/app/src/app_state/stage_orchestration.rs
@@ -40,7 +40,7 @@ impl AppState {
 
         self.update_camera_uniforms();
 
-        self.renderer.update_view_rect_overlay(
+        self.renderer.update_selection_rect_overlay(
             &self.graphic.queue,
             self.active_selection_rect,
             self.graphic.config.width,
