+++ view/app/src/lib.rs
@@ -7,9 +7,9 @@ pub mod graphic;
 pub mod logging;
 pub mod mouse_input;
 pub mod overlay_coords;
+pub mod selection_rect;
+pub mod selection_rect_renderer;
 pub mod snapshot_overlay_renderer;
 pub mod stage_factory;
 pub mod stl_loader;
 pub mod svg_loader;
-pub mod view_rect;
-pub mod view_rect_renderer;
