+++ view/app/src/selection_rect.rs
